feat:피드백 반영 및 Test 분리
