+++ src/test/java/com/gamzabat/algohub/feature/studygroup/controller/StudyGroupControllerTest.java
@@ -4,11 +4,9 @@ import static org.hamcrest.Matchers.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.doNothing;
 import static org.mockito.BDDMockito.doThrow;
-import static org.mockito.BDDMockito.eq;
 import static org.mockito.BDDMockito.times;
 import static org.mockito.BDDMockito.verify;
 import static org.mockito.BDDMockito.when;
-import static org.mockito.BDDMockito.*;
 import static org.mockito.Mockito.anyLong;
 import static org.mockito.Mockito.anyString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -40,11 +38,9 @@ import com.gamzabat.algohub.common.DateFormatUtil;
 import com.gamzabat.algohub.common.jwt.TokenProvider;
 import com.gamzabat.algohub.config.SpringSecurityConfig;
 import com.gamzabat.algohub.enums.GroupStatus;
-import com.gamzabat.algohub.enums.JoinRequestStatus;
 import com.gamzabat.algohub.exception.StudyGroupValidationException;
 import com.gamzabat.algohub.exception.UserValidationException;
 import com.gamzabat.algohub.feature.group.studygroup.controller.StudyGroupController;
-import com.gamzabat.algohub.feature.group.studygroup.domain.JoinRequest;
 import com.gamzabat.algohub.feature.group.studygroup.dto.BookmarkStatus;
 import com.gamzabat.algohub.feature.group.studygroup.dto.CheckSolvedProblemResponse;
 import com.gamzabat.algohub.feature.group.studygroup.dto.CreateGroupRequest;
@@ -58,11 +54,9 @@ import com.gamzabat.algohub.feature.group.studygroup.dto.GroupCodeResponse;
 import com.gamzabat.algohub.feature.group.studygroup.dto.GroupRoleResponse;
 import com.gamzabat.algohub.feature.group.studygroup.dto.UpdateBookmarkResponse;
 import com.gamzabat.algohub.feature.group.studygroup.dto.UpdateGroupMemberRoleRequest;
-import com.gamzabat.algohub.feature.group.studygroup.dto.UpdateJoinRequestStatusRequest;
 import com.gamzabat.algohub.feature.group.studygroup.etc.RoleOfGroupMember;
 import com.gamzabat.algohub.feature.group.studygroup.exception.CannotFoundGroupException;
 import com.gamzabat.algohub.feature.group.studygroup.exception.GroupMemberValidationException;
-import com.gamzabat.algohub.feature.group.studygroup.exception.JoinRequestException;
 import com.gamzabat.algohub.feature.group.studygroup.repository.GroupMemberRepository;
 import com.gamzabat.algohub.feature.group.studygroup.repository.StudyGroupRepository;
 import com.gamzabat.algohub.feature.group.studygroup.service.JoinRequestService;
@@ -877,119 +871,6 @@ class StudyGroupControllerTest {
 			.andExpect(status().isForbidden())
 			.andExpect(jsonPath("$.error").value("참여하지 않은 그룹입니다."));
 		verify(studyGroupService, times(1)).editStudyGroupVisibility(user, groupId, request);
-	}
-
-	@Test
-	@DisplayName("가입 요청 성공")
-	void joinRequestSuccess() throws Exception {
-		//given
-		Long groupId = 10L;
-		willDoNothing().given(joinRequestService).joinRequest(any(User.class), eq(groupId));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request", groupId)
-				.header("Authorization", token))
-			.andExpect(status().isOk());
-	}
-
-	@Test
-	@DisplayName("가입 요청 실패 : 이미 요청한 그룹")
-	void joinRequest_fail_alreadyRequested() throws Exception {
-		Long groupId = 10L;
-		willThrow(new JoinRequestException("이미 요청한 그룹입니다."))
-			.given(joinRequestService).joinRequest(any(User.class), eq(groupId));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request", groupId)
-				.header("Authorization", token))
-			.andExpect(status().isBadRequest());
-	}
-
-	@Test
-	@DisplayName("가입 요청 목록 조회")
-	void getAllJoinRequests_success() throws Exception {
-		Long groupId = 10L;
-		// 직렬화가 비어도 배열 길이만 확인할 수 있게 더미 객체 2개
-		given(joinRequestService.getAllJoinRequests(any(User.class), eq(groupId)))
-			.willReturn(List.of(new JoinRequest(), new JoinRequest()));
-
-		mockMvc.perform(get("/api/groups/{groupId}/join-request", groupId)
-				.header("Authorization", token))
-			.andExpect(status().isOk())
-			.andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-			.andExpect(jsonPath("$", hasSize(2)));
-	}
-
-	@Test
-	@DisplayName("요청 목록 조회 실패 : 권한 없음")
-	void getAllJoinRequests_fail_alreadyRequested() throws Exception {
-		Long groupId = 10L;
-		given(joinRequestService.getAllJoinRequests(any(User.class), eq(groupId)))
-			.willThrow(new JoinRequestException("요청 목록을 조회할 권한이 없습니다."));
-
-		mockMvc.perform(get("/api/groups/{groupId}/join-request", groupId)
-				.header("Authorization", token))
-			.andExpect(status().isBadRequest());
-	}
-
-	@Test
-	@DisplayName("승인 성공")
-	void approve_success() throws Exception {
-		Long groupId = 10L;
-		Long requestId = 77L;
-		UpdateJoinRequestStatusRequest request = new UpdateJoinRequestStatusRequest(JoinRequestStatus.APPROVE);
-		willDoNothing().given(joinRequestService)
-			.updateJoinRequest(any(User.class), eq(requestId), eq(groupId), eq(request));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request/{requestId}", groupId, requestId)
-				.header("Authorization", token)
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(request)))
-			.andExpect(status().isOk());
-	}
-
-	@Test
-	@DisplayName("승인 실패 해당 요청 없음")
-	void approve_fail_requestNotFound() throws Exception {
-		Long groupId = 10L;
-		Long requestId = 999L;
-		UpdateJoinRequestStatusRequest request = new UpdateJoinRequestStatusRequest(JoinRequestStatus.APPROVE);
-		willThrow(new JoinRequestException("해당 요청이 존재하지 않습니다"))
-			.given(joinRequestService).updateJoinRequest(any(User.class), eq(requestId), eq(groupId), eq(request));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request/{requestId}", groupId, requestId)
-				.header("Authorization", token))
-			.andExpect(status().isBadRequest());
-	}
-
-	@Test
-	@DisplayName("가입 요청 거절 성공")
-	void reject_success() throws Exception {
-		Long groupId = 10L;
-		Long requestId = 77L;
-		UpdateJoinRequestStatusRequest request = new UpdateJoinRequestStatusRequest(JoinRequestStatus.REJECT);
-
-		willDoNothing().given(joinRequestService)
-			.updateJoinRequest(any(User.class), eq(requestId), eq(groupId), eq(request));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request/{requestId}", groupId, requestId)
-				.header("Authorization", token)
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(request)))
-			.andExpect(status().isOk());
-	}
-
-	@Test
-	@DisplayName("가입 요청 거절 실패 권한 없음")
-	void reject_fail_noPermission() throws Exception {
-		Long groupId = 10L;
-		Long requestId = 77L;
-		UpdateJoinRequestStatusRequest request = new UpdateJoinRequestStatusRequest(JoinRequestStatus.REJECT);
-
-		willThrow(new JoinRequestException("승인 권한이 없습니다."))
-			.given(joinRequestService).updateJoinRequest(any(User.class), eq(requestId), eq(groupId), eq(request));
-
-		mockMvc.perform(post("/api/groups/{groupId}/join-request/{requestId}", groupId, requestId)
-				.header("Authorization", token))
-			.andExpect(status().isBadRequest());
 	}
 
 }
